requirements.txt
click
inquirerpy
pyyaml
reprint